requirements.txt
black==23.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
idna==3.4
isort==5.12.0
mypy==1.4.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.7.0
PyYAML==6.0.1
requests==2.31.0
termcolor==2.3.0
toml==0.10.2
tomli==2.0.1
types-requests==2.31.0.1
types-toml==0.10.8.6
types-urllib3==1.26.25.13
typing_extensions==4.6.3
urllib3==2.0.3
xmltodict==0.13.0
